+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 1. Improve Table's client function handleRemoveDuplicate for much faster rendering.
 2. Add closeCursor to statement in PdoDataSource.
-3. Add `Shuffle` process to randomize data order.
+3. Fix the string function with null parameter in PHP8.1
+4. Upgrade jquery to 3.5.0
+5. Add `Shuffle` process to randomize data order.
 
 ## Version 5.6.2